Refactors JavaScript to be more in line with Magento conventions and simplifies build process due to capabilities within Magento to handle part of this.
