+++ view/frontend/requirejs-config.js
@@ -1,6 +1,6 @@
 var config = {
   paths: {
-    magento_plugin: 'AddressFinder_AddressFinder/js/magento-plugin-min',
+    magento_plugin: 'AddressFinder_AddressFinder/js/magento-plugin',
     addressfinder: 'https://api.addressfinder.io/assets/v3/widget'
   },
 };
