+++ view/frontend/web/js/source/address-form-config/shipping-checkout.js
@@ -1,5 +1,5 @@
 
-import stateMappings from './state_mappings'
+import stateMappings from './state-mappings'
 
 export default {
   label: "Shipping Checkout",
